The Project of CALCULATOR, is been made using Django <for backend> and HTML<frontend>.
Talking about DJANGO, its a module<inbuilt packages> of PYTHON <3.0 and higher versions>; which is used to code for the backend side of a static or dynamic website.
Secondly , for the front end part , here we have used the HTML language<with basic tags>.
This projects particularly foreshadows the working DJANGO module in backend of a website.
<IMAGES OF THE WEBSITE ARE AVAILABLE FOR THE OUTPUT REFERENCE>
That's all from myside.
 Thank you for reading.